common: Add new functions to CMDU client
Preparative commit.
Add Two new functions to the CmduClient class that are required on the
for the Agent thread:

1. forward_cmdu() to forward a cmdu_rx to a server socket.
It is now required to forward messages from the fronthauls thread to the
backhaul manager thread.

2. get_fd() to get the file descriptor of a CMDU client. It is required
to be able to compare the file descriptor of which the message was
received on to a member class that holds it (e.g m_ap_manager_fd), so we
could identify from which socket the message has been received and
handled it properly.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>