agent: Add EventLoop components
Add EventLoop components to the agent thread:
1. BrokerClient - to send/receive a message to/from the BUS.
2. CmduServer - to accept incoming connections
3. TimerManager - to run periodic FSM function.
4. Replace the platform and the backhaul type from SocketClient to
CmduClient.
5. Replaced the monitor and the ap_manger type from Socket to
FileDescriptor (int).

Also, add functions that provide the capability to send/forward CMDU to
a client or the broker.
For the broker, also add a handler function.

Initialized all new object in the thread_init() function.

Use the new backhaul and platform manager client objects to open the
connection.
Note that on the object event handlers we use handle_cmdu() and
agent_reset(), which will be introduced in a future commit.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>