agent: Replace master_socket with link_to_controller()
Since the member "master_socket" has been removed in a previous commit,
add a function link_to_controller() that reflects the same
functionality:
Check if there is a link to the Controller by checking if the
auto-configuration
has successfully finished. This indicates that the Controller has
been discovered.

Add a clearing of the AgentDB autoconfiguration completion status on
the Agent INIT state.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>