agent: Replace the use of sockets to FileDescriptors and EventLoop
Until now, the agent used sockets and sockets function for connectivity.
In the previous commits, all sockets were replaced with:
FileDescriptors, CmduClient, and added broker interface which has
connectivity APIs of its own, all based on EventLoop.

The class server infrastructure was replaced with the CmduServer which
is also based on EventLoop.

Replace across the Agent file (son_slave) the use of the old socket with
the new EventLoop based solutions.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>