agent : Remove IRE_MAX_WIRELESS_RECONNECTION_TIME_MSC
IRE_MAX_WIRELESS_RECONNECTION_TIME_MSC timer setting is redundant.
It is set only in a single place, and not being checked.
On backhaul disconnect, we do agent reset which ends up going to the
INIT state. This means that setting the timer for backhaul reconnection
is irrelevant.
I guess we previously had a retry mechanism that has been removed and
this is code is probably a leftover of it.
Remove IRE_MAX_WIRELESS_RECONNECTION_TIME_MSC completely.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>