controller: refactor report handling in db
Currently to store the DCSr2 returning reports we store said reports in
a vector of wfa_map::cNeighbors. This causes issues as the
wfa_map::cNeighbors object holds only pointers to a buffer containing
the result parameters.
To correctly handle the results, we need to store them in a container
that holds the information internally.

Additionally, the DCSr2 currently uses MID values to align requests
with incoming reports, this is incorrect.

To correctly handle the reports we need an indexing system that will
allow us to identify which incoming report belongs to which outgoing
request.

- node: add channel_scan_report_records to align split reports with
a timestamp.
- node: use beerocks_message::sChannelScanResults instead of
wfa_map::cNeighbors to store incoming reports.
- db: add get_channel_scan_report to return a scan report for a given
radio
- db: add get_channel_scan_report to return a scan report for a given
radio and index
- db: add get_channel_scan_report to return a scan report for a given
radio and timestamp
- db: refactor add_channel_report to store incoming reports in a
container of beerocks_message::sChannelScanResults and update the
related index.

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>