DCSr2 task: Fix channel list in controller request
Currently there is an unknown issue that causes the supported channel
list in the controller's DB to contain invalid channels.

Because the DCS uses the supported channels list for it's channel pool,
the channel pool could also contains invalid channels.

When creating the channel scan request in the controller task, we need
to only store valid central channels.

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>