DCSr2 task: refactor handle_scan_report_event
As part of the effort to refactor the DCSr2 from using MID values to
using timestamps we need to change how the controller handles the
finalized DCSr2 results.

Refactor the handle_scan_report_event function to rely on timestamps.

When finalizing the ongoing scan requests, use the reported timestamp
to validate the matching report index in the DB to the report index
expected by the request.

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>