agent: Add DCSr2 support for zero dwell-time.
To support the zero dwell-time feature in the DCSr2 we need the agent
task to send a dump request instead of a trigger request.

Doing so will result in a scan-less dump of the cached results from the
monitor.

- tlvf: Add support for BACKHAUL_CHANNEL_SCAN_DUMP_RESULTS_REQUEST.
- son_slave: handle incoming BACKHAUL_CHANNEL_SCAN_DUMP_RESULTS_REQUEST
messages and forward them to the monitor.
- channel_scan_task: differentiate between a dump request and a scan
request and send the correct request forward.
- channel_scan_task: handle incoming CHANNEL_SCAN_DUMP_RESULTS_RESPONSE
messages.

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>