controller: add default gateway nodes/objects
In case of "Controller Only Mode", beerocks agent will not trigger
SLAVE JOIN action.
It will be kept as dummy application to run tasks and transport layer.

In that scenario, we need to add gateway objects manually to make it
possible to add extenders to network.
Otherwise, gateway is not found and network can not be generated.

https://jira.prplfoundation.org/browse/PPM-1562
--
Signed-off-by: Cihangir Odabas <codabas@assia-inc.com>