bpl: Add new BPL functions
On the newly added TLV "tlvDeviceInventory" we are required to fill
fields:
1. "Serial Number" - A string Identifying the particular device that is
unique for the indicated model and manufacturer.
2. "Chipset Vendor" - A string identifying the Wi-Fi chip vendor of this
radio.

Both fields have information that is not "configuration" therefore less
fit to UCI which contain configuration only, but moreover, these pieces
of information could not be defined in UCI during compilation time,
since they shall be the same for several HW models.
Add a new folder in BPL called "vendor" which contains functions that
their implementation should be overridden by the vendor.
Add functions to get both new pieces of information required by the new
TLV, and implement it to return some dummy value so we can pass
certification.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>