bwl: nl80211: don't empty association frames
In b020c19d63c77225e94558ede5985928aac3f999, we added support for
receiving, storing and finally sending association frames when we
receive STA_Connected messages. A check was included to make sure we
only store association frames from the time we receive them, to the
time we receive the AP-STA-CONNECTED event.

This worked fine for regular clients but for 4addr-mode stations,
other messages are received in-between which makes the AP manager
stop:
 event received:<3>AP-MGMT-FRAME-RECEIVED buf=[]
Saved assoc frame
 event received:<3>WDS-STA-INTERFACE-ADDED ifname=wlan0.sta1 sta_addr=[]
Unhandled event received: WDS-STA-INTERFACE-ADDED
 event received:<3>AP-STA-CONNECTED []
STA-CONNECTED without previously receiving a (re-)association frame!

We didn't notice this issue before in certification (e.g. in
MAP-4.2.3) because:

- in certification, fronthaul and backhaul are separate BSS.

- events from other BSSes are not received by the AP
manager (PPM-1714).

To fix it, don't remove the stored association frame when we receive
other messages.

Ideally we should replace it by a check that parses the association
frame to make sure the MAC address involved is the same as the one
from the STA_Connected event. There is however some work ongoing to
fix the association frames parsing (see [1]), so for now just add a
TODO with a linked Jira task.

The check for an empty association frame has been kept, but turned
into a warning.

Fixes PPM-1715

[1]: https://gitlab.com/prpl-foundation/prplmesh/prplMesh/-/merge_requests/2682

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>