backhaul_manager: re-instate the code for already connected bSTA
The code that handled the case where the bSTA is already connected when
prplMesh starts up was removed a couple of years ago as a quick fix to
get certification to work. Re-instate it as a quick fix to get wireless
backhaul to work.

The state machine move is inspired on what happens when we receive the
Connected event.