prplmesh_utils.sh: remove ebtables rule
The ebtables rule causes a problem for wireless backhaul: when the
wireless backhaul is established before prplmesh starts, prplmesh
doesn't add the wlan0.sta1 interface to the list of interfaces that the
transport has to monitor. Therefore, it won't receive any packets that
arrive on the interface instead of on the bridge.

By removing the ebtables rule, all the packets will arrive on the bridge
instead of on the interface, and the transport does monitor the bridge
so it receives the multicast packets.

Note that LLDP packets will still not be received, because those are
dropped by the kernel itself (hardcoded rule).