agent: Simplify FSM
Preparative commit.
If the Autoconfiguration task is on the Agent context, all of the radio
specific states could be moved inside the task since these states are
part of the autoconfiguration flow.

Removed Autoconfiguration states from the Agent FSM.
The STOP and the OPERATIONAL states are now part of the Agent FSM.
The PRE_OPERATIONAL state is no longer needed since the OPERATIONAL
state is being done exactly once now.
The states VERSION_MISMATCH and SSID_MISMATCH, were removed since the
STATE_STOPPED covers them anyway.
The stop flag in the radio context moved to the Agent context.
Radio states were removed completely.
As a result of all removed states, the slave_fsm() function removed
entirely since it does not have any state left in it.

There are still some references to things that were removed on the
slave_thread file.They will be moved to the autoconfiguration task
on a future commit and will be changed then.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>