bwl: nl80211: remove max_num_sta setting
When the re-implementation of update_vap_credentials() was copied from
dwpal, the setting of max_num_sta was included. However, that's a stupid
limitation of the dwpal drivers, most drivers have no reason to limit
the number of stations that can connect (and if they do, they'll apply
the limit themselves, no need to configure hostapd for it).

Remove the setting of max_num_sta, so more than one backhaul station can
connect.