bwl: nl80211: fix out of bound access of association frame
If get_binary_association_frame is given a zero-length C array, it's
going to access it out of bounds.

Fix it by setting association_frame_length to 0 (which is what the AP
manager will check for) and not using get_binary_association_frame if
there is no association frame to be included.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>