Revert "prplmesh_utils.sh: remove ebtables rule"
This reverts commit 34b84d5481708388f49bef9df1b9e6ae72de22e8.

Without the ebtables rules, some messages were flooded (the 1905
multicast messages most likely) which lead to the connection between
devices to be completely unusable.
When this happened, the following was visible in the system logs:
Fri Oct 29 13:37:54 2021 kern.warn kernel: [12784.463561] br-lan: received packet on wlan0.sta4 with own address as source address (addr:b2:83:c4:0d:f4:ae, vlan:0)

We initially tried to come up with another ebtables rule that would
only drop some of the traffic, but it didn't work.

After a lot of local testing, we found out that keeping the ebtables
rule was not a problem, at least if the devices are started in a
specific order (controller first, then the agents one by one).  With
the ebtables rule enabled again, we were able to connect 5 devices to
the controller were prplMesh was operational on all of them, and the
connection was still stable. Note that due to a mistake when testing,
it was tested with max_num_sta set to 1 (so it's still possible that
the ebtables is in the end still needed when multiple STAs are
connected to the same AP).