DCSr2: removing spamming log in the DCSr2 task
In the DCSr2 controller task, when checking if agent is idle with any
pending radio scans, we return false if there aren't any pending scans.

Since this is checked almost every iteration of the DCSr2 controller
task. and it should usually return false in that condition, there is no
need to print a log.

Remove the spamming log in DCSr2 controller task.

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>