DCSr2: create new agent handler for continuous scan
When creating a new continuous scan request we prematurely create an
agent status handler before knowing if the scan is valid or not, this
could result in an error where the agent status handler is created but
since the scan is invalid (for example due to an invalid dwell-time) no
scan element is added, thus resulting in an "empty" agent status
handler.

Create the agent element after passing the scan validation.

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>