controller: remove some of the tasks
The controller has performance issues where it has troubles onboarding
devices when the total number of devices starts to grow.

When this happens, the controller process fully uses one core (25% CPU
usage on a GL.iNet B1300).

We were not able to identify which part of the controller is causing
it, but on the other hand there are a number of tasks and message
handlers that we don't need for the demo.

Just disable some them for now, to try to reduce the load on the
controller.

With theses changes, we were at least able to manually onboard 11
external agents (so 12 devices were part of the network in total).