controller: don't remove dead nodes
One issue we currently face is that one of the agents gets rejected:
INFO 10:46:05:798 <3069818208> controller.cpp[2205] --> IRE Slave
joined
    slave_version=2.2.1
    slave_name=IRE_54:9c
    parent_bssid_mac=96:83:c4:11:56:d2
    backhaul_mac=94:83:c4:11:54:9c
    backhaul_ipv4=192.168.1.178
    bridge_mac=b2:83:c4:11:54:9c
    bridge_ipv4=192.168.1.178
    backhaul_manager=0
    backhaul_type=WIFI_INTEL
    low_pass_filter_on = 0
    radio_mac = 96:83:c4:11:54:9d
    channel = 0
    is_gw_slave = 0

DEBUG 10:46:05:799 <3069818208> controller.cpp[2231] --> sending back
join reject!
DEBUG 10:46:05:799 <3069818208> controller.cpp[2232] --> reject_debug:
parent_bssid_has_node=1

When checking on the agent '96:83:c4:11:56:d2', we can see that it was
onboarded.

Also, 'parent_bssid_has_node=1' indicates that the node is the
database, but probably not in the CONNECTED state.

What possibly happens is that at some point this agent is not in one
of the topology messages, so it gets removed. Later on it gets added
in the database again, but not in the CONNECTED state.

To avoid this issue for the demo, just don't remove "dead" nodes.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>