ci: platform: replace netgear-rax40 with nec-wx3000hp
The upcoming version of MaxLinear's Wifi driver will not be compatible
with the Netgear rax40, so this device will be deprecated soon.

Remove it from the release artifacts.

In the boardfarm tests, replace it with the NEC WX3000HP.
While at it, also take the chance to use $TARGET_DEVICE_NAME instead
of hard-coding the device name.

In the .test-on-target job, add the NEC WX3000HP to ALL_TARGETS to
make sure we stop prplMesh on it, for it not to interfere with other
tests.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>