config_file: add my_var to beerocks controller configuration
my_var is a new parameter that is introduced to the controller config file.
The expected values of my_var range is uint32's range. 1 is the default value.

This commit adds the parameter support to the config file, and the following commits will
add the read parameter to the controller DB and add set/get capabilities through CLI commands.

Signed-off-by: Yoni Abecassis <yoniab93@gmail.com>