tests: Fix bug on V2Certification_4_7_10 boardfarm test
On this test, there is a construction of METRIC_REPORTING_POLICY TLV.
After the place holder for the radio mac, there are 5 octets.
When looking at the TLV documentation, it is possible to see that after
the radio mac, there is a place only for four octets.
Remove the last octet assignment from the TLV.

Note: It is not clear how the test has passed until now.
Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>