tests: Fix bug on PolicyConfigWMetricReportingPolicy boardfarm test
On this test, there is a call to the function
send_and_check_policy_config_metric_reporting().
This argument requires passing Controller and Agent objects as
arguments in order of controller first, and agent as a second.
On this specific test, the arguments are provided in the opposite order.
Fix it by flipping the order of the arguments.

Note: It is not clear how the test has passed until now.
Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>