bcl: Add const qualifiers after functions
Preparative commit.
On a future commit, new functions that return a CmduClient object are
going to be added to the Agent class, to allow external entities (tasks)
send messages.
We wouldn't want that an external entity will be able to modify the
internal properties of the object, to do it we must return the object as
a constant.
To do that, add 'const' qualifier to some base function on base classes.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>