agent: Add tasks to the Agent and move there the Autoconfiguration task
Add to the Agent the ability to run tasks by adding a task pool to the
Agent, and initializing the tasks timer.

In addition, move the AP-Autoconfiguration task initialization from the
Backhaul Manager to the Agent.
As part of it, move the tasks required messages subscriptions from
the Backhaul Manager to the Agent, and allow tasks to be able to handle
messages coming from the transport.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>