agent: Prepare the Agent with new functions
Tasks will need to use some of the Agent class members and functions.
Some of them are private and we wouldn't want to expose them bare to
external tasks.
Add getters function that allows external task entities to get access to
the Agent class members as constants.
In addition, modify the functions that send CMDUs to the Controller, to
support multicast.

On the function send_cmdu_to_controller() remove the Controller link
validation, since the validation is implemented by checking the
Auto-Configuration status, but since this function will be needed before
the controller is being discovered, it is expected to fail the function,
therefore the check has been removed.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>