agent: Move all helper functions
Move all helpers functions used by the Auto-Configuration flow from the
Agent (son_slave) file to the Auto-Configuration task file.

Notice that for MULTI_AP_POLICY_CONFIG_REQUEST_MESSAGE there were two
handlers, one on the Link Metrics task where it was originally handled
and has been forwarded to the Agent.
The reason for this design is that on R1 this message contained only
Link Metrics policy, but on R2 new TLVs unrelated to the Link Metrics
were added. Since this message is more a configuration message rather
than a Link Metrics message, all of its handlings moved to the
Auto-Configuration task. To do that, add policy configuration fields
on the AgentDB.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>