agent: Change FSMs in the Agent and the Backhaul Manager
Since now the full the Auto-Configuration flow is being done in
the task, remove all the related code leftovers from the Agent FSM, and
thus remove completely the radio specific states.
Moved the configuration of the transport from the Backhaul Manager,
to the Agent. Now after the Agent receives "Backhaul Connected"
notification in trigger the Auto-configuration task by sending it an
START_AP_AUTOCONFIGURATION event and move to
STATE_WAIT_FOR_AUTO_CONFIGURATION_COMPLETE. Once the auto-configuration
is complete, the Agent will move to an OPERATIONAL state.

On the Backhaul Manager side, remove the controller discovery states,
so now after the backhaul is connected, it will report it to the Agent
and move to OPERATIONAL state right away.

Note that the Topology task AGENT_DEVICE_INITIALIZED event handler has
been changed to update the periodic discovery message sending instead of
sending the discovery message. This is since, then, the Backhaul Manager
triggered that event after finished the auto-configuration has
completed, but now it is right after establishing the backhaul link and
before the transport is configured by the auto-configuration task in the
Agent.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>