agent: Refactor the Search message response handler
We are sending the search message multiple times to support different
Profiles Controllers. As a result, we might get several responses.
If we already completed the discovery return right away to prevent
redundant processing and log printing.

Move the parsing of beerocks handshake message after the check on the
discovery status since it is redundant to do it if the Controller
already been discovered.

In addition, add the missing parsing of the MultiApProfile TLV and set
the Controller Profile on the AgentDB as it was missing too.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>