bwl: reduce rate of generate of client connected events
At prplmesh start, the BWL generates STA_Connected events for
pre-connected clients so they will be part of the controller DB and can
be monitored.

In cases where there are many clients, the generate time (which uses
dwpal commands to access hostapd to get the information of the clients)
can take very long time in which the monitor/ap-manager can be
non-responsive to agent requests.

Reduce the rate to 5 msec each iteration and 500 msec between
iterations.

PPM-1703.

(ported from WLANRTSYS-31891).

Signed-off-by: Adam Dov <adov@maxlinear.com>