controller: nbapi: set correct last connect time of station
LastConnectTime is the time after station is associated.
Whenever it is read, it needs to be updated.

TimeStamp object is used before to calculate this,
but it is used to track when statistics is received.

So, it is now calculated from assoc. timestamp which is available in db.

Time handling is converted to cpp style.

https://jira.prplfoundation.org/browse/PPM-1470
--
Signed-off-by: Cihangir Odabas <codabas@assia-inc.com>