controller: skip processing old M1 messages
In some situations, the controller is so overloaded that it takes a
lot of time to process each message it receives.

When this happens, M2 messages can be sent more than 5 seconds (the
timeout of the M1 message on the agent) after the corresponding M1
message was received.

Since the cryptographic operations the controller has to do for each
M2 messages are quite expensive, skip those messages (since the
receiving agent will anyway not be able to use the M2 messages after
the M1 timeout).

Fixes PPM-1783

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>