tests: environment.py commands as list of args
Some inconsistency was found when sending commands to both a container
device ('docker exec') and a real device ('ssh'). Docker commands were
being sent as list of executable args, whereas a real device command
were sent as a single string.
This would make a scenario with a function declared on the ALEntity
class something not viable for one of the child classes.
After a vote on the democracy channel it was decided that the best
option was to keep every command as a list of executable commands.

ALEntityRDKB was copy-pasted from ALEntityPrplWrt before 030b6823
was introduced. When changes in 030b6823 were released they
were not added to the ALEntityRDKB. This commit brings 030b6823
changes to the ALEntityRDKB.

Signed-off-by: Dmytro Puz <d.puz@inango-systems.com>