prplOS: bump to include downgraded netifd for MaxLinear
The netIFD version used by MaxLinear is older than the version used by
prplOS. The introduction of 8.5.2 raised issues related to the version
difference.

To resolve the issues, there is a need to introduce netIFD patches but
these patches from MaxLinear cannot be easily ported to the new netIFD.
Therefore, the decision is to use a downgraded version for MaxLinear
platforms (aligned with the netIFD version used internally by
MaxLinear) and to port security patches introduced by prplOS to that
feed.

This commit introduces the new separate feed for netIFD.

PPM-1778.

Signed-off-by: Adam Dov <adov@maxlinear.com>