bcl: check if ebtables's entry exist before removing it
Adding or removing ebtables entry is done through
an os_utils::system_call function with the specified arguments.
When removing ebtables's entry request is issued, the function above
will try to remove an unexisting entry, resulting in an error log.
To avoid an error log, we check beforehand whether this entry exists,
and if no, we don't issue removing command.

Signed-off-by: Yoni Abecassis <yabecassis@maxlinear.com>