agent: fix AP_AUTOCONFIGURATION_WSC_MESSAGE CMDU handling
AP_AUTOCONFIGURATION_WSC_MESSAGE is a message that is sent
from both agent and controller, and both processes handles
this message upon receiving one from the other side.
The agent sends this message with tlvWsc TLV that contains M1,
and the controller sends this message with tlVwsc TLV that contains M2.
However, since the agent is subscribed to this message, it will receive
both the M2 message from the controller, and the M1 message
that it sent to the controller.
Thus, the agent shall not handle the message
that is intended to the controller.

Signed-off-by: Yoni Abecassis <yabecassis@maxlinear.com>