cmake: populate monitor_polling_rate_msec param at compile time
With current default monitor_polling_rate ie 250ms on UGW and RDKB platform,
it is seen that monitors' CPU utilization and load on hostapd is high.
Experiment values shows 1000ms fits well for UGW and RDKB platforms.

So, set default value for monitor_polling_rate_msec as 250msec and
1000ms for UGW and RDKB platform at compile time.

PPM-1856

Signed-off-by: anantkumar <cchavda@maxlinear.com>