agent: Improve Topology Notification performance
According to IEEE_Std_1905.1-2013 specification, if some information
specified in the Topology Response message has changed, then within one
second, a Topology Notification must be sent.
If it happens a lot of times in one second period, it could lead to
a flood of Topology Notifications on the network, and to a system slow
down, since Topology Notification is a 'reliable multicast' message,
which means, it also needs to be sent as unicast to all the 1905.1
neighbors.
To prevent such a situation, we limit sending more than one Topology
Notification in a second.

Add a mechanism that when the function send_topology_notification()
is called, then a timer of one second starts. On timeout the Topology
Notification will be sent when called from the task work() function.

In addition, on the Topology Discovery handler function, use the class
'send_topology_notification()' function which has the added mechanism,
instead of creating and sending the message manually. It also prevents
code duplication.

While at it, replace iterator increment from 'it++' form to '++it'
which is faster.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>