transport: break forwarding loop from redundant iterations
When the transport forwards an outgoing message to the bridge or to
the network interfaces, it enters a loop which iterates on all the
monitored interfaces and check if need to send the message to the
currently iterated interface.
When sending a unicast message, this loop will send the packet only to
the bridge, and continue to loop on all other interfaces as well
although it is redundant.
Add a brake condition that stops the loop if the destination is the
the bridge, and the packet already forwarded to it.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>