transport: Add missing update of src_if_index
When updating the 'last_seen' time of a 1905 neighbor, we don't update
the src_if_index. This is a bug since the interface of which we
received a message from that neighbor might change and the src_if_index
will be wrong.
Update the src_if_index every time we receive a message from a neighbor.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>