ci: boardfarm-on-nec-wx3000hp: add specific tags
We now have a second testbed that we can use for Boardfarm tests.

Up to now we used the "targets" tag to run tests on targets, and this
included both the Boardfarm tests on NEC WX3000HP and the operational
tests on Turris Omnia and GL.iNet.

The new testbed currently only has a NEC WX3000HP, so it can't be used
to run the operational tests.

To be able to choose to run Boardfarm tests only, per runner, add a
new "boardfarm" tag. An additional tag, "nec-wc3000hp", is also added
so that in the future if some Boardfarm testbeds can run tests on some
devices but not others, we still have the flexibility to configure
which device is available for each runner.

While at it, remove the line that changed backhaul_wire_iface in the
agent configuration file, as PPM-230 is complete and the wired
backhaul interface is configured in UCI instead.

Fixes PPM-1467

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>