controller: fix retrieving of RDKB wlan task ID
The controller uses DB for storing the task ID of RDKB wlan task.
This task ID is saved in DB when the task pool executes
rdkb_wlan_task's work() method.
Meanwhile, the controller tries to access this task ID from the DB
from another place, but since the task pool has not scheduled rdkb task
yet, the task ID is not saved in the DB yet.
The solution is to assign task ID in rdkb_wlan_task constructor,
instead of in work() method, in order to avoid this synchronization issue.
In addition, the state variable in rdkb_wlan_task class is redundant.

Signed-off-by: Yoni Abecassis <yabecassis@maxlinear.com>