bml: Prevent possible thread hang
If the calling to the thread_base::stop() from outside the thread
context, it can lead to a race with the thread_init() function that
runs within the thread context.
To prevent the race, keep the same flow of the thread_base::run()
function but without changing the 'should_stop' flag. Thus the race
is prevented.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>