agent: Filter multiple Agent reset
When something in the system goes bad, we call 'agent_reset()',
the problem is that when such a thing happens, usually many things will
go bad and 'agent_reset()' will be called from multiple locations.
Add a state filter that checks if the Agent state is on an initial
state, if so, don't continue with the 'agent_reset()' function.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>