agent: Don't stop the fronthaul process on agent_reset()
Currently, if we call agent_reset(), we stop all the fronthauls.
This behavior is unwanted. If for instance one of the fronthauls crash,
we would not want to stop the fronthaul that is OK.
Removed the call to fronthaul_stop() on the agent_reset() function.
If some of the fronthauls crash, it will be handled by a socket
disconnection handler.
Note that also the initialization of the radio mac has been removed,
since it does not have any purpose now.

While at it, simplify the agent_reset(), by removing unnecessary
branching and nestings, so it will have the same behavior but more
readable.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>