agent: Disconnect the other fronthaul thread
If a fronthaul stops for some reason, both ap_manager and monitor will
disconnect. The order of the events is random.
If we get the event on one of them, disconnect the other one.
Also, re-initialize the socket file descriptor to invalid.
The reason for that is that we need to make sure that if one of the
fronthaul threads is down/disconnected then all of them will be, so the
part of the boot sequence that stops and starts the fronthaul process
will do that only on fronthaul who its threads marked as invalid.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>