agent: Disconnect ap_manager instead of agent_reset
Until now if there was an error related to the fronthaul we would do
"agent_reset" which killed both fronthauls and start them again.
Now the agent_reset() does not kill the fronthaul anymore, it is only
reset the agent FSM.
Some errors do require reset to the Fronthaul process.
To do that now, replace the call to agent_reset() to ap_manager
disconnect. It will eventually end up with agent_reset with both
fronthaul threads file descriptors marked as invalid, so the fronthaul
process will be restarted.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>