agent: Clear channel list only for non-active fronthauls
On the Agent STATE_INIT, we are clearing the channels list of the AP.
If the fronthaul is already up we don't want to clear it since the AP is
sending it in the JOIN message only once, so if we clear it, we will not
receive it again.
On the Agent STATE_INIT clear the channels list only if the AP Manager
file descriptor is invalid.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>