framework: bpl_cfg: disable wlan3 for turris omnia
PrplMesh code assumes that board has 3 WLAN ifaces but there are
only "wifi0" and "wifi1" on TurrisOmnia. And if 3rd agent for "wifi2"
is present and continiously try to start then prplMesh cannot become
Operational.

IRE_MAX_SLAVES is equal 2 for TurrisOmnia and prevents starting agent for
"wifi2" WLAN iface.

Signed-off-by: Dmytro Puz <d.puz@inango-systems.com>