common: Add custom command implementation
Add the implementation of the custom command on the UCC Listener derived
classes.
Agent implementation is a stub for now.
On the Controller add the implementation of the 'discovery_burst'
command.
Note that there is an emphasis on sending all the messages in a very
short time. That is why we just edit certain parameters before sending
the buffer again rather than create a new CMDU that internally does
a memset of 4096 bytes to zero.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>