tests: Add topology_discovery_burst to the borardfarm test
Add a topology_discovery_burst test to boardfarm the command the
Controller to send 214 discovery messages as being done on R3 4.7.10_ETH
certification test, and checks that processing in the Agent of the last
message has been done in less than 0.5 seconds.
Add the test to the testsuits.cfg file.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>