ci: don't create release-related jobs for non-release pipelines
In commit f2f9999538cd6b038c9e9ea5073782cee307a23e, the rules for
creating release-related jobs (changelog, upload-release-artifacts and
release) have been changed: while only being created when part of a
release pipeline before, they are now created unconditionally, but
either started automatically if part of a release, or manually if not.

When a manual job depends on another manual job that hasn't been
started yet, the whole pipeline is marked as "blocked". Because of
this, we cannot merge MRs anymore (see PPM-1877).

The original intention behind this change was to be able to run
release-related jobs manually outside of releases to test them, so:

- For the changelog job, running it outside of release is possible, so
keep its behavior as-is by copying the "onrelease" anchor.

- For the upload-release-artifacts, we cannot currently run it outside
of releases because '$CI_COMMIT_TAG' is generally not set. We could
use another name to upload packages to the registry, but this is made
more difficult by the fact that the package version is limited to a
strict regex (see [1]). Instead, make it so that the job is not
created outside of releases. This means that to test the job when
developing, its rules will have to be changed on a side branch.

- For the release job itself, it's even more complicated. Since Gitlab
currently considers the whole pipeline blocked because the manual
release job depends on other manual jobs, we cannot keep it as a
manual job at all. The way we worked-around it in the past for the
certification tests was to move them to a child pipeline (the
"platform" pipeline). Since this would be too much to do just for the
release job, do like for 'upload-release-artifacts' and don't create
the job when not part of a release (the change is done in the
"onrelease" YAML anchor now shared only by the
"upload-release-artifacts" and the "release" jobs).

Fixes PPM-1877

[1]: https://docs.gitlab.com/ee/user/packages/generic_packages/#publish-a-package-file

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>