agent: Fix AP attach failure leads to crash
In a real platform, prplmesh might be up before the HOSTAP is ready for
attachment. When that happens, the AP Manager will detach right away and,
the agent will get a disconnect event which causes the Agent FSM to
restart.
From the Agent FSM point of view, it will happen before the Monitor
joins, thus since the Monitor socket file descriptor is not set, it also
resets the Fronthaul of the other radio too.
Eventually, it would lead to a loop that each time the Agent gets to the
JOIN_INIT state, it will stop the other fronthaul and its socket
disconnect handler will restart the Agent again.

To fix it, replace the condition of stop/start the fronthauls, to be
based on a flag that indicates whether the Fronthaul started or not.
If it started, we will not stop and start the Fronthaul, instead, we will
let it complete its join sequence.

In addition, to preserve the behavior we had where the backhaul link is
disconnected then the Fronthauls goes down, add a condition to the Agent
reset function to stop the fronthaul if the backhaul is disconnected.

To prevent a situation when an AP Manager disconnects before it sent the
UP_NOTIFICATION and the Agent crash, change the LOG level from FATAL to
INFO.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>