link_metrics_collection_task: get_neighbor_links: Filter neighbor per interface name
When multiple interfaces have the same MAC address, any 1905 neihgbor
connected to one of them is reported as connected to each of them, thus
leading to duplication and incoherent data.
This issue comes from the filtering being done only on the interface MAC
address, which appears insufficient in this case.

Filtering also on the interface name prevents such issue. The interface
name must be collected and stored along the interface MAC address to
make this filtering possible

Signed-off-by: Sagemcom <prplmesh.opensources@sagemcom.com>