ApMetrics TLV: Collect channel utilization even if no threshold has been defined
When no Channel Utilization Threshold was defined, this stat was never
retrieved from the HAL layer and thus the value set in the ApMetrics TLV
was always 0. However, the Multi-AP spec doesn't request a threshold to
be defined to report the actual value of this stat. Only the
Agent-initiated sending of an AP Metrics Response frame must be
conditionned to the definition of such threshold.

Signed-off-by: Sylvain LE ROUX <prplmesh.opensources@sagemcom.com>